+++ example/lib/pages/home.dart
@@ -1,8 +1,0 @@
-import 'package:get_server/get_server.dart';
-
-class HomePage extends GetView {
-  @override
-  build(Context context) {
-    return Text('Hello, you are on home');
-  }
-}