+++ example/lib/pages/html/html.dart
@@ -1,0 +1,11 @@
+import 'dart:io';
+
+import 'package:get_server/get_server.dart';
+
+class HtmlPage extends GetView {
+  @override
+  build(Context context) {
+    final path = '${Directory.current.path}/example/web/index.html';
+    return Html(path);
+  }
+}