+++ example/lib/pages/json/json.dart
@@ -1,0 +1,10 @@
+import 'package:get_server/get_server.dart';
+
+class JsonPage extends GetView {
+  @override
+  build(Context context) {
+    return Json({
+      "fruits": ["banana", "apple", "orange"]
+    });
+  }
+}