+++ example/lib/pages/upload/upload.dart
@@ -1,0 +1,16 @@
+import 'dart:convert';
+
+import 'package:get_server/get_server.dart';
+
+class UploadPage extends GetView {
+  @override
+  Future<Widget> build(Context context) async {
+    final upload = await context.file('file');
+    final data = {
+      "nameFile": upload.name,
+      "mimeType": upload.mimeType,
+      "fileBase64": "${base64Encode(upload.data)}",
+    };
+    return Json(data);
+  }
+}