+++ example/lib/pages/user/data/data_provider/user_data.dart
@@ -1,0 +1,18 @@
+abstract class IUserProvider {
+  Map<String, dynamic> getUser(String name);
+}
+
+class UserProvider extends IUserProvider {
+  Map<String, dynamic> getUser(String name) {
+    if (name.toLowerCase() == 'pedro') {
+      return {
+        'name': 'Pedro',
+        'age': '30 years old',
+        'Country': 'Brazil',
+        'Error': ''
+      };
+    } else {
+      return {'Error': 'User Not found'};
+    }
+  }
+}