+++ lib/get_server.dart
@@ -2,3 +2,6 @@ export 'src/context/context_request.dart';
 export 'src/core/server_main.dart';
 export 'src/routes/route.dart';
 export 'src/widgets/pageable.dart';
+export 'src/infrastructure/getx_controller.dart';
+export 'package:get_instance/get_instance.dart';
+export 'package:get_rx/get_rx.dart';