+++ lib/src/logger/log.dart
@@ -1,5 +1,0 @@
-typedef LogWriterCallback = void Function(String text, {bool isError});
-
-void logger(String value, {bool isError = false}) {
-  print(value);
-}